backend/requirements.txt
flask
flask-cors
requests
packaging
google-generativeai
web3
python-dotenv
